// solutis-agile-frontend/src/api/generated/types/SupplierCreateIn.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { AddressPayload } from './AddressPayload.ts'
import type { CompanyInformationPayload } from './CompanyInformationPayload.ts'
import type { ContactPayload } from './ContactPayload.ts'
import type { ContractPayload } from './ContractPayload.ts'
import type { FiscalDetailsPayload } from './FiscalDetailsPayload.ts'
import type { OrganizationalDetailsPayload } from './OrganizationalDetailsPayload.ts'
import type { PaymentDetailsPayload } from './PaymentDetailsPayload.ts'

/**
 * SupplierCreateIn
 * @description Supplier creation payload.
 */
export type SupplierCreateIn = {
  /**
   * @type string
   */
  legalName: string
  /**
   * @type string
   */
  taxId: string
  tradeName?: string | null
  stateBusinessRegistration?: string | null
  municipalBusinessRegistration?: string | null
  classification?: number | null
  category?: number | null
  riskLevel?: number | null
  type?: number | null
  address?: AddressPayload | null
  contact?: ContactPayload | null
  paymentDetails?: PaymentDetailsPayload | null
  organizationalDetails?: OrganizationalDetailsPayload | null
  fiscalDetails?: FiscalDetailsPayload | null
  companyInformation?: CompanyInformationPayload | null
  contract?: ContractPayload | null
}