// solutis-agile-frontend/src/api/generated/types/SupplierOut.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { AddressOut } from './AddressOut.ts'
import type { CompanyInformationOut } from './CompanyInformationOut.ts'
import type { ContactOut } from './ContactOut.ts'
import type { ContractOut } from './ContractOut.ts'
import type { DomainRefOut } from './DomainRefOut.ts'
import type { FiscalDetailsOut } from './FiscalDetailsOut.ts'
import type { OrganizationalDetailsOut } from './OrganizationalDetailsOut.ts'
import type { PaymentDetailsOut } from './PaymentDetailsOut.ts'
import type { SupplierSituationOut } from './SupplierSituationOut.ts'

/**
 * SupplierOut
 * @description Supplier detail response.
 */
export type SupplierOut = {
  /**
   * @type integer
   */
  id: number
  name?: string | null
  tradeName?: string | null
  /**
   * @type string
   */
  legalName: string
  /**
   * @type string
   */
  taxId: string
  stateBusinessRegistration?: string | null
  municipalBusinessRegistration?: string | null
  address?: AddressOut | null
  contact?: ContactOut | null
  paymentDetails?: PaymentDetailsOut | null
  organizationalDetails?: OrganizationalDetailsOut | null
  fiscalDetails?: FiscalDetailsOut | null
  companyInformation?: CompanyInformationOut | null
  contract?: ContractOut | null
  classification?: DomainRefOut | null
  category?: DomainRefOut | null
  riskLevel?: DomainRefOut | null
  type?: DomainRefOut | null
  createdAt?: string | null
  updatedAt?: string | null
  situation?: SupplierSituationOut | null
  responsibilityMatrix?: {
    [key: string]: any
  } | null
}