// solutis-agile-frontend/src/api/generated/types/SupplierUpdateIn.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { AddressPayload } from './AddressPayload.ts'
import type { CompanyInformationPayload } from './CompanyInformationPayload.ts'
import type { ContactPayload } from './ContactPayload.ts'
import type { ContractPayload } from './ContractPayload.ts'
import type { FiscalDetailsPayload } from './FiscalDetailsPayload.ts'
import type { OrganizationalDetailsPayload } from './OrganizationalDetailsPayload.ts'
import type { PaymentDetailsPayload } from './PaymentDetailsPayload.ts'

/**
 * SupplierUpdateIn
 * @description Supplier update payload.
 */
export type SupplierUpdateIn = {
  legalName?: string | null
  taxId?: string | null
  tradeName?: string | null
  stateBusinessRegistration?: string | null
  municipalBusinessRegistration?: string | null
  classification?: number | null
  category?: number | null
  riskLevel?: number | null
  type?: number | null
  address?: AddressPayload | null
  contact?: ContactPayload | null
  paymentDetails?: PaymentDetailsPayload | null
  organizationalDetails?: OrganizationalDetailsPayload | null
  fiscalDetails?: FiscalDetailsPayload | null
  companyInformation?: CompanyInformationPayload | null
  contract?: ContractPayload | null
}